requirements.txt
django~=4.2.0
djangorestframework~=3.14.0
django-cors-headers~=4.2.0
psutil~=5.9.0
requests~=2.31.0
channels==4.2.0